+++ packages/components/organisms/f-cookie-banner/src/tenants/da-DK.js
@@ -5,7 +5,8 @@ const messages = {
     acceptButtonText: 'Accepter alle cookies',
     textLine1: 'Vi anvender de cookies (og lignende teknologier), der er nødvendige, for at platformen kan fungere.',
     textLine2: 'Vi vil også gerne bruge valgfri cookies til forbedring af platformen ved at indsamle oplysninger om, hvordan du bruger den.',
-    textLine3: 'For nærmere oplysninger om de anvendte cookies og teknologier henviser vi til vores {cookiePolicy}. Ved brug af denne banner placeres en cookie på din enhed, som husker dine præferencer.',
+    textLine3: 'For nærmere oplysninger om de anvendte cookies og teknologier henviser vi til vores ',
+    textLine4: '. Ved brug af denne banner placeres en cookie på din enhed, som husker dine præferencer.',
     cookiePolicyLinkText: 'cookie-meddelelse',
     cookiePolicyLinkUrl: 'https://www.just-eat.dk/info/cookiepolitik'
 };