+++ packages/components/organisms/f-cookie-banner/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.3.0
+------------------------------
+*February 25, 2021*
+
+### Changed
+- Update tests to use js-cookie.
+- Update legacy cookie value
+- Remove f-globalisation mixin
+
 v0.2.0
 ------------------------------
 *February 17, 2021*